+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - `Ctrl/Cmd + d` shortcut in About window to show debug information
 - ability to hide dock icon on macOS (requires restart)
+- monitoring suspend/resume (alfa, electron has bugs)
 
 ### Fixed
 - break is not skippable via shortcut in strict mode